requirements.txt
discord.py>=2.3.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
webdavclient3>=3.14.6 